Handle cases where SCRIPT_NAME/URL are empty
Prefix SCRIPT_NAME even if it's empty, but ignore SCRIPT_URL if it's
empty.

This is because SCRIPT_URL isn't required, but if it exists, it's
supposed to include the script name *and* the path info. So if it's
empty and we add a prefix and the request wasn't for the root path, the
SCRIPT_URL will be wrong. And since django prefers SCRIPT_URL over
SCRIPT_NAME when it exists, we'll get bad paths.

The heart of this change is that it makes this stuff work when we're
behind apache, where SCRIPT_URL is likely to be set (and preferred by
django), and in local dev, where SCRIPT_URL *shouldn't* be set (and
therefore *should* be ignored by django).